need to fix error messages, hd in pipes and echo redirects

+++ srcs/exec/pipex.c
@@ -6,7 +6,7 @@
 /*   By: ibenhaim <ibenhaim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/31 21:46:28 by ibenhaim          #+#    #+#             */
-/*   Updated: 2023/09/25 13:41:09 by ibenhaim         ###   ########.fr       */
+/*   Updated: 2023/09/25 15:20:52 by ibenhaim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,11 +36,8 @@ void	redir_pipes(t_data *data, t_piplist *cur)
 	}
 	ft_close(&cur->redir.sstdin);
 	switch_and_close_fds(data);
-	if (cur->redir.in)
-	{
-		if (cur->redir.hd && access(cur->redir.in, F_OK) == 0)
-			unlink(cur->redir.in);
-	}
+	if (cur->redir.in && cur->redir.hd && access(cur->redir.in, F_OK) == 0)
+		unlink(cur->redir.in);
 }
 
 void	pipex_end(t_piplist *cur, t_piplist *cpy, t_data *data)
